+++ christmas_app/models.py
@@ -10,7 +10,7 @@ from . import db
 class User(db.Model, UserMixin, AnonymousUserMixin):
     """ Database table: user
         each user account setting/properties defined here.
-        
+
         #Attribute:
             fname -> first name,
             alias -> alias (not null),
@@ -18,12 +18,11 @@ class User(db.Model, UserMixin, AnonymousUserMixin):
             points -> foreign key to Game table
         """
     id = db.Column(db.Integer(), unique=True, primary_key=True)
-    fname = db.Column(db.String(56), unique=True) # first name
+    fname = db.Column(db.String(56), unique=True)  # first name
     alias = db.Column(db.String(20), nullable=False)
     password = db.Column(db.String())
     points = db.relationship('Game')
     langPref = db.Column(db.String(2), default="EN")
-
 
     def __str__(self):
         return self.fname
@@ -33,7 +32,7 @@ class User(db.Model, UserMixin, AnonymousUserMixin):
         if not current_user.is_authenticated:
             # return current_app.login_manager.unauthorized()
             return False
-        elif current_user.is_authenticated and (current_user.fname == 'KITTIPICH') :
+        elif current_user.is_authenticated and (current_user.fname == 'KITTIPICH'):
             return True
         return False
 
@@ -57,14 +56,14 @@ class Game(db.Model):
     finish_time = db.Column(db.DateTime())
     score = db.Column(db.Integer(), nullable=False, default=0)
     counter = db.Column(db.Integer(), nullable=False, default=0)
-    played_by = db.Column(db.String(), db.ForeignKey('user.fname'), nullable=False, default="Guest")
-
+    played_by = db.Column(db.String(), db.ForeignKey(
+        'user.fname'), nullable=False, default="Guest")
 
     def __str__(self) -> str:
         return str("ID: " + self.id + " Finished at: " + self.finish_time)
 
-    def __init__(self, finish_at:datetime.datetime, score:int,played_by,start_at:datetime.datetime=None):
-        def increment() -> int: # increment doesn't work, need to be fixed
+    def __init__(self, finish_at: datetime.datetime, score: int, played_by, start_at: datetime.datetime = None):
+        def increment() -> int:  # increment doesn't work, need to be fixed
             current = self.counter
             if current == None:
                 current = 0
@@ -77,12 +76,10 @@ class Game(db.Model):
         self.counter = increment()
         self.played_by = played_by
 
-
-
-class Question(db.Model):
-    """ Database table: question
-        each question that will ask users
-        """
-    q_id = db.Column(db.Integer(), nullable=False, primary_key=True) # question id
-    q = db.Column(db.String(), nullable=False) # question
-    a = db.Column(db.String(), nullable=True) # answer
+# class Question(db.Model):
+#     """ Database table: question
+#         each question that will ask users
+#         """
+#     q_id = db.Column(db.Integer(), nullable=False, primary_key=True) # question id
+#     q = db.Column(db.String(), nullable=False) # question
+#     a = db.Column(db.String(), nullable=True) # answer